+++ libdh/math.c
@@ -4,8 +4,8 @@
 
 #include "../include/dhmath.h"
 
-const double PI_dh = 3.14159265358979323846;
-const double E_dh = 2.7182818284590452354;
+const long double PI_dh = 3.14159265358979323846;
+const long double E_dh = 2.7182818284590452354;
 
 double factorial_dh(double a){
 	double retval = 1.0;
